scripts/requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
apt-clone==0.2.1
apturl==0.5.2
asttokens==2.4.1
async-timeout==4.0.1
attrs==25.3.0
Babel==2.8.0
bcrypt==3.2.0
beautifulsoup4==4.13.4
blinker==1.4
blis==1.3.0
blurhash==1.1.4
Brlapi==0.8.3
Brotli==1.0.9
catalogue==2.0.10
certifi==2025.6.15
chardet==4.0.0
charset-normalizer==3.4.2
chromedriver-binary==74.0.3729.6.0
click==8.2.1
cloudpathlib==0.21.1
colorama==0.4.4
comm==0.2.2
command-not-found==0.3
commonmark==0.9.1
confection==0.1.5
configobj==5.0.6
contourpy==1.3.2
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
cymem==2.0.11
dbus-python==1.2.18
debugpy==1.8.5
decorator==5.2.1
defer==1.0.6
distro==1.7.0
exceptiongroup==1.3.0
executing==2.0.1
eyeD3==0.8.10
filelock==3.6.0
fonttools==4.58.4
frozenlist==1.4.1
greenlet==3.2.3
gyp==0.1
h11==0.16.0
httpcore==1.0.5
httplib2==0.20.2
httpx==0.27.0
i==2.1.0
idna==3.10
ifaddr==0.1.7
IMDbPY==2021.4.18
importlib-metadata==4.6.4
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.6
jiter==0.5.0
joblib==1.5.1
jupyter_client==8.6.2
jupyter_core==5.7.2
keyring==23.5.0
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
marisa-trie==1.2.1
Markdown==3.3.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
Mastodon.py==2.0.1
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==8.10.0
multidict==6.0.5
murmurhash==1.0.13
mutagen==1.45.1
nemo-emblems==6.0.1
nest-asyncio==1.6.0
netaddr==0.8.0
netifaces==0.11.0
nltk==3.9.1
numpy==2.2.6
oauthlib==3.3.1
onboard==1.4.1
openai==1.40.6
outcome==1.3.0.post0
packaging==25.0
PAM==0.4.2
pandas==2.3.0
parso==0.8.4
pexpect==4.8.0
pillow==11.3.0
platformdirs==4.2.2
pnpm==1.0.14
praw==7.8.1
prawcore==2.4.0
preshed==3.0.10
prompt_toolkit==3.0.47
protobuf==3.12.4
proton-core==0.4.0
proton-keyring-linux==0.2.0
proton-vpn-api-core==0.42.5
proton-vpn-gtk-app==4.9.7
proton-vpn-network-manager==0.12.13
psutil==5.9.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycairo==1.20.1
pycryptodomex==3.11.0
pycups==2.0.1
pycurl==7.44.1
pydantic==2.11.7
pydantic_core==2.33.2
pyelftools==0.27
Pygments==2.19.2
PyGObject==3.42.1
PyICU==2.8.1
pyinotify==0.9.6
PyJWT==2.3.0
pymacaroons==0.13.0
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==3.2.3
pyparted==3.11.7
PyPDF2==3.0.1
pyRFC3339==1.1
PySocks==1.7.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.1.1
python-gnupg==0.4.8
python-magic==0.4.27
python-xlib==0.29
pytz==2025.2
pyxdg==0.27
PyYAML==5.4.1
pyzmq==26.1.0
qrcode==7.3.1
regex==2024.11.6
reportlab==3.6.8
requests==2.32.4
requests-file==1.5.1
requests-oauthlib==2.0.0
rglob==1.7
rich==14.0.0
seaborn==0.13.2
SecretStorage==3.3.1
selenium==4.34.0
sentry-sdk==1.4.3
setproctitle==1.2.2
shellingham==1.5.4
six==1.17.0
smart_open==7.3.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.41
srsly==2.5.1
ssh-import-id==5.11
stack-data==0.6.3
style==1.1.0
systemd-python==234
textblob==0.19.0
thinc==8.3.6
tinycss2==1.1.1
tldextract==3.1.2
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
trio==0.30.0
trio-websocket==0.12.2
tweepy==4.16.0
tweety==0.1.6
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
ubuntu-drivers-common==0.0.0
ufw==0.36.1
Unidecode==1.3.3
update==0.0.1
update-checker==0.18.0
urllib3==2.4.0
vaderSentiment==3.3.2
vboxapi==1.0
wadllib==1.3.6
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==9.1
wordcloud==1.9.4
wrapt==1.17.2
wsproto==1.2.0
xdg==5
xkit==0.0.0
xlrd==1.2.0
yarl==1.9.4
yt-dlp==2022.4.8
zipp==1.0.0
